refactor(block-header): Added a spearate block_test.go for validating whether the correct corresponding function is executed in NewBlockHeaderFromCbor.
